Created new folder data with file "nav.ts" with array of elements end states for language button
