+++ zoo-hope/src/App.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import logo from './images/icons/logo.svg';
+import { NavBar } from './components/navBar';
 
 function App() {
   useEffect(() => {
@@ -12,20 +13,7 @@ function App() {
 
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+      <NavBar/>
     </div>
   );
 }
